requirements.txt
Flask
langchain
langchain-community
faiss-cpu
openai
werkzeug
requests
python-dotenv
